backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
bcrypt>=3.2.0 # Explicitly listing bcrypt, passlib will use it
pydantic
pydantic-settings
python-dotenv
aiosqlite
# For real payment integrations (examples, choose as needed):
# stripe
# paypalrestsdk
# coinbase-commerce (or other crypto libraries)